chip-list uses transformer to get array of objects from array of strings
